+++ settings.gradle.kts
@@ -39,6 +39,7 @@ include(":app-android")
 include(":app-shared")
 include(":shared")
 include(":shared-core")
+include(":shared-domain")
 include(":shared-ui")
 
 include(":internal-ktlint")
