+++ shared-domain/api/shared-domain.api
@@ -1,0 +1,58 @@
+public final class dev/sdkforge/camera/domain/CameraConfig {
+	public fun <init> ()V
+	public fun <init> (Ljava/util/Set;ZZLdev/sdkforge/camera/domain/Facing;)V
+	public synthetic fun <init> (Ljava/util/Set;ZZLdev/sdkforge/camera/domain/Facing;ILkotlin/jvm/internal/DefaultConstructorMarker;)V
+	public final fun component1 ()Ljava/util/Set;
+	public final fun component2 ()Z
+	public final fun component3 ()Z
+	public final fun component4 ()Ldev/sdkforge/camera/domain/Facing;
+	public final fun copy (Ljava/util/Set;ZZLdev/sdkforge/camera/domain/Facing;)Ldev/sdkforge/camera/domain/CameraConfig;
+	public static synthetic fun copy$default (Ldev/sdkforge/camera/domain/CameraConfig;Ljava/util/Set;ZZLdev/sdkforge/camera/domain/Facing;ILjava/lang/Object;)Ldev/sdkforge/camera/domain/CameraConfig;
+	public fun equals (Ljava/lang/Object;)Z
+	public final fun getCameraFacing ()Ldev/sdkforge/camera/domain/Facing;
+	public final fun getEnableTorch ()Z
+	public final fun getFormats ()Ljava/util/Set;
+	public final fun getUseHapticFeedback ()Z
+	public fun hashCode ()I
+	public fun toString ()Ljava/lang/String;
+}
+
+public final class dev/sdkforge/camera/domain/Facing : java/lang/Enum {
+	public static final field BACK Ldev/sdkforge/camera/domain/Facing;
+	public static final field FRONT Ldev/sdkforge/camera/domain/Facing;
+	public static fun getEntries ()Lkotlin/enums/EnumEntries;
+	public static fun valueOf (Ljava/lang/String;)Ldev/sdkforge/camera/domain/Facing;
+	public static fun values ()[Ldev/sdkforge/camera/domain/Facing;
+}
+
+public final class dev/sdkforge/camera/domain/Format : java/lang/Enum {
+	public static final field AZTEC Ldev/sdkforge/camera/domain/Format;
+	public static final field CODE_128 Ldev/sdkforge/camera/domain/Format;
+	public static final field CODE_39 Ldev/sdkforge/camera/domain/Format;
+	public static final field CODE_93 Ldev/sdkforge/camera/domain/Format;
+	public static final field DATA_MATRIX Ldev/sdkforge/camera/domain/Format;
+	public static final field EAN_13 Ldev/sdkforge/camera/domain/Format;
+	public static final field EAN_8 Ldev/sdkforge/camera/domain/Format;
+	public static final field ITF Ldev/sdkforge/camera/domain/Format;
+	public static final field PDF417 Ldev/sdkforge/camera/domain/Format;
+	public static final field QR_CODE Ldev/sdkforge/camera/domain/Format;
+	public static final field UNKNOWN Ldev/sdkforge/camera/domain/Format;
+	public static final field UPC_E Ldev/sdkforge/camera/domain/Format;
+	public static fun getEntries ()Lkotlin/enums/EnumEntries;
+	public static fun valueOf (Ljava/lang/String;)Ldev/sdkforge/camera/domain/Format;
+	public static fun values ()[Ldev/sdkforge/camera/domain/Format;
+}
+
+public final class dev/sdkforge/camera/domain/ScanResult {
+	public fun <init> (Ldev/sdkforge/camera/domain/Format;Ljava/lang/String;)V
+	public final fun component1 ()Ldev/sdkforge/camera/domain/Format;
+	public final fun component2 ()Ljava/lang/String;
+	public final fun copy (Ldev/sdkforge/camera/domain/Format;Ljava/lang/String;)Ldev/sdkforge/camera/domain/ScanResult;
+	public static synthetic fun copy$default (Ldev/sdkforge/camera/domain/ScanResult;Ldev/sdkforge/camera/domain/Format;Ljava/lang/String;ILjava/lang/Object;)Ldev/sdkforge/camera/domain/ScanResult;
+	public fun equals (Ljava/lang/Object;)Z
+	public final fun getFormat ()Ldev/sdkforge/camera/domain/Format;
+	public final fun getValue ()Ljava/lang/String;
+	public fun hashCode ()I
+	public fun toString ()Ljava/lang/String;
+}
+
