+++ shared-domain/build.gradle.kts
@@ -1,0 +1,28 @@
+plugins {
+    alias(libs.plugins.kotlinMultiplatform)
+    alias(libs.plugins.androidLibrary)
+    alias(libs.plugins.binaryCompatibilityValidator)
+    alias(libs.plugins.dokka)
+    alias(libs.plugins.build.logic.library.kmp)
+    alias(libs.plugins.build.logic.library.android)
+    alias(libs.plugins.build.logic.library.publishing)
+}
+
+kotlin {
+    sourceSets {
+        commonMain {
+            dependencies {
+                // put your multiplatform dependencies here
+            }
+        }
+        commonTest {
+            dependencies {
+                implementation(libs.kotlin.test)
+            }
+        }
+    }
+}
+
+android {
+    namespace = "dev.sdkforge.camera.domain"
+}
