+++ shared-domain/src/commonMain/kotlin/dev/sdkforge/camera/domain/Format.kt
@@ -1,4 +1,4 @@
-package dev.sdkforge.camera.ui
+package dev.sdkforge.camera.domain
 
 enum class Format {
     AZTEC,
