+++ shared-domain/src/commonMain/kotlin/dev/sdkforge/camera/domain/ScanResult.kt
@@ -1,6 +1,6 @@
 @file:Suppress("ktlint:standard:class-signature")
 
-package dev.sdkforge.camera.ui
+package dev.sdkforge.camera.domain
 
 data class ScanResult(
     val format: Format,
