+++ shared-ui/src/androidMain/kotlin/dev/sdkforge/camera/ui/CameraView.android.kt
@@ -9,6 +9,7 @@ import androidx.compose.ui.platform.LocalHapticFeedback
 import androidx.compose.ui.viewinterop.AndroidView
 import androidx.lifecycle.compose.LifecycleResumeEffect
 import androidx.lifecycle.compose.LocalLifecycleOwner
+import dev.sdkforge.camera.domain.CameraConfig
 
 @Composable
 actual fun CameraView(
