+++ shared-ui/src/androidMain/kotlin/dev/sdkforge/camera/ui/Format.android.kt
@@ -1,5 +1,6 @@
 package dev.sdkforge.camera.ui
 
+import dev.sdkforge.camera.domain.Format
 import com.google.mlkit.vision.barcode.common.Barcode as MLBarcode
 
 internal val Format.value: Int
