+++ shared-ui/src/androidMain/kotlin/dev/sdkforge/camera/ui/NativeCameraControllerImpl.android.kt
@@ -3,6 +3,7 @@ package dev.sdkforge.camera.ui
 import android.content.Context
 import androidx.compose.ui.hapticfeedback.HapticFeedback
 import androidx.lifecycle.LifecycleOwner
+import dev.sdkforge.camera.domain.CameraConfig
 
 internal actual class NativeCameraControllerImpl(
     context: Context,
