+++ shared-ui/src/commonMain/kotlin/dev/sdkforge/camera/ui/CameraView.kt
@@ -2,6 +2,7 @@ package dev.sdkforge.camera.ui
 
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
+import dev.sdkforge.camera.domain.CameraConfig
 
 @Composable
 expect fun CameraView(
@@ -13,10 +14,3 @@ expect fun CameraView(
 expect fun rememberCameraController(
     config: CameraConfig = CameraConfig(),
 ): CameraController
-
-data class CameraConfig(
-    val formats: Set<Format> = Format.entries.toSet(),
-    val enableTorch: Boolean = false,
-    val useHapticFeedback: Boolean = false,
-    val cameraFacing: Facing = Facing.BACK,
-)
